Work in progress to try to add music to game
Sort of really hard because SadConsole isn't built with audio in mind. It doesn't really use the MonoGame content pipeline at all - which creates a ton of issues in regards to building the audio assets.

Feels silly not to use monogame in the project, because monogame 'is' what SadConsole is built on top of, just really hard to utilize audio in sadconsole at all.